+++ packages/lib/src/markdown-shims.d.ts
@@ -28,3 +28,8 @@ declare module "markdown-it-source-map" {
   const plugin: PluginSimple;
   export default plugin;
 }
+declare module "markdown-it-deflist" {
+  import type { PluginSimple } from "markdown-it";
+  const plugin: PluginSimple;
+  export default plugin;
+}